Code Quality: Merge namespaces that don't need to be separated into namespaces for the same purpose
