--- src/Files.App/Controllers/IJson.cs
@@ -1,9 +1,0 @@
-namespace Files.App.Controllers
-{
-	internal interface IJson
-	{
-		string JsonFileName { get; }
-
-		void SaveModel();
-	}
-}
