+++ src/Files.App/Converters/StringToBrushConverter.cs
@@ -3,7 +3,7 @@ using Microsoft.UI.Xaml.Data;
 using Microsoft.UI.Xaml.Media;
 using System;
 
-namespace Files.App.ValueConverters
+namespace Files.App.Converters
 {
 	internal sealed class StringToBrushConverter : IValueConverter
 	{
@@ -12,7 +12,7 @@ namespace Files.App.ValueConverters
 			if (value is not string strValue)
 				return null;
 
-			return new SolidColorBrush(ColorHelper.ToColor(strValue));
+			return new SolidColorBrush(strValue.ToColor());
 		}
 
 		public object ConvertBack(object value, Type targetType, object parameter, string language)
