+++ src/Files.App/Views/WidgetsPage.xaml.cs
@@ -5,7 +5,6 @@ using Files.App.Filesystem;
 using Files.App.Helpers;
 using Files.App.UserControls.Widgets;
 using Files.App.ViewModels;
-using Files.App.ViewModels.Pages;
 using Files.Backend.Services.Settings;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Navigation;
